Fixes for Copilots suggestions part2, API

+++ api-gateway/src/main/java/com/petclinic/bffapigateway/dtos/Bills/BillResponseDTO.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -22,9 +23,9 @@ public class BillResponseDTO {
     private String vetFirstName;
     private String vetLastName;
     private LocalDate date;
-    private double amount;
-    private double taxedAmount;
-    private double interestAmount;
+    private BigDecimal amount;
+    private BigDecimal taxedAmount;
+    private BigDecimal interestAmount;
     private BillStatus billStatus;
     private LocalDate dueDate;
     private Long timeRemaining;
